bugfix (mem and fat32)

+++ kernel/syscall/sys/types.h
@@ -1,15 +1,15 @@
 #ifndef _SYS_TYPES
 #define _SYS_TYPES
 
-typedef unsigned int    dev_t;
-typedef unsigned int    ino_t;
-typedef unsigned short  mode_t;
+typedef short           dev_t;
+typedef unsigned short  ino_t;
+typedef unsigned int    mode_t;
 typedef unsigned short  nlink_t;
-typedef unsigned int    uid_t;
-typedef unsigned int    gid_t;
-typedef unsigned int    off_t;
-typedef unsigned int    blksize_t;
-typedef unsigned int    blkcnt_t;
-typedef unsigned int    time_t;
+typedef unsigned short  uid_t;
+typedef unsigned short  gid_t;
+typedef long            off_t;
+typedef long            blksize_t;
+typedef long            blkcnt_t;
+typedef long            time_t;
 
 #endif
